services/scraper/pyproject.toml
[tool.poetry]
name = "eventing-scraper"
version = "0.1.0"
description = "Scraper for eventing"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "*"
beautifulsoup4 = "*"
python-dateutil = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
